Add ruff check (#2262)
* Updates for numpy 2.0 release; Add ruff check; Cleanup tests

* Degeneracy issue in KGKS tests

* Adjust tests

* Fix a knumint bug when handling mcol and ncol

* Update ruff config

+++ pyscf/mp/ump2.py
@@ -303,7 +303,7 @@ def make_fno(mp, thresh=1e-6, pct_occ=None, nvir_act=None, t2=None, eris=None):
     masks = _mo_splitter(mp)
 
     if nvir_act is not None:
-        if isinstance(nvir_act, (int,numpy.int)):
+        if isinstance(nvir_act, (int, numpy.integer)):
             nvir_act = [nvir_act]*2
 
     no_frozen = []
